clutter: Change stage view scale to be float
To support fractional scaling, change the stage view scale to be a
float instead of an int. Also change the places where it is retrieved
and used when scaling things.

+++ clutter/clutter/clutter-actor.c
@@ -17708,7 +17708,7 @@ _clutter_actor_calculate_resource_scale (ClutterActor *self,
   ClutterActor *stage;
   GList *views;
   GList *l;
-  int max_scale = 0;
+  float max_scale = 0;
 
   stage = _clutter_actor_get_stage_internal (self);
   if (!stage)
